fix #14

+++ build/server/lib/repo.js
@@ -71,20 +71,24 @@ module.exports.create = function(app, callback) {
 
 module.exports["delete"] = function(app, callback) {
   var child;
+  console.log(app.userDir);
   child = spawn('rm', ['-rf', app.userDir]);
   return child.on('exit', function(code) {
     if (code !== 0) {
       return callback(new Error('Unable to remove directory'));
     } else {
-      return fs.unlink(app.logFile, function(err) {
-        if (fs.existsSync(app.backup)) {
-          return fs.unlink(app.backup, function(err) {
+      console.log(app.logFile);
+      if (fs.existsSync(app.logFile)) {
+        return fs.unlink(app.logFile, function(err) {
+          if (fs.existsSync(app.backup)) {
+            return fs.unlink(app.backup, function(err) {
+              return callback();
+            });
+          } else {
             return callback();
-          });
-        } else {
-          return callback();
-        }
-      });
+          }
+        });
+      }
     }
   });
 };
